src/main/Singleton.java
package main;

import java.nio.charset.StandardCharsets;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.Arrays;

public class Singleton {

    private static Singleton singleton = null;

    private Singleton() {
    }

    public static Singleton getInstance() {
        if (singleton == null)
            singleton = new Singleton();
        return singleton;
    }

    public String generateUID(Publicacao pub) throws NoSuchAlgorithmException {
        MessageDigest digest = MessageDigest.getInstance("SHA-256");
        byte[] hash = digest.digest((pub.getEvento() + pub.getProcesso() + pub.getTexto()).getBytes(StandardCharsets.UTF_8));
        return Arrays.toString(hash);
    }
}


